// register/types.ts
export interface RegisterFormData {
  nome: string;
  email: string;
  telefone: string;
  cpf: string;
  dataNascimento: string;
  pais: string;
  estado: string;
  cidade: string;
  cep: string;
  rua: string;
  numero: string;
  apartamento: string;
  logradouro: string;
  senha: string;
  confirmarSenha: string;
}